Update library to support python 3.12 and pandas>=2.0
This patch fixes a unescaped string symbols, like "\s", which are
raising SyntaxError in Python 3.12 by replacing strings with raw
strings.

Also patch changes series iteration `iteritems()`, which was removed
in Pandas>=2.0 and replaces it with `iloc` iterator, which is available
in both, Pandas<2.0 and Pandas>=2.0

Patch also bumps version of the library to 0.3.7.